completed repo to price his and editing wire

+++ migrations/000004_create_price_history.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS price_history (
     id               SERIAL PRIMARY KEY,
     game_id          INTEGER NOT NULL REFERENCES games(id) ON DELETE CASCADE,
     lang             VARCHAR(2) NOT NULL,
-    price            INTEGER NOT NULL,
+    price            NUMERIC(10, 2) NOT NULL,
     currency         VARCHAR(3) NOT NULL,
     discount_percent INTEGER NOT NULL DEFAULT 0,
     checked_at       TIMESTAMP NOT NULL DEFAULT NOW()
